Refactor item/segment queries into services
