+++ src/components/header/CommandPalette.tsx
@@ -19,7 +19,7 @@ import { Film, Loader2, Mic2, Play, Search, Tv, X } from 'lucide-react'
 import type { BaseItemDto } from '@/types/jellyfin'
 import { Dialog, DialogContent } from '@/components/ui/dialog'
 import { Button } from '@/components/ui/button'
-import { useItems } from '@/hooks/queries/use-items'
+import { useItems } from '@/services/items/queries'
 import { useSelectedCollectionSearch } from '@/hooks/use-selected-collection-search'
 import { useVirtualWindow } from '@/hooks/use-virtual-window'
 import { cn } from '@/lib/utils'
