+++ src/routes/series/$itemId.tsx
@@ -7,18 +7,16 @@ import { Suspense, lazy } from 'react'
 import { createFileRoute, notFound } from '@tanstack/react-router'
 import { z } from 'zod'
 
-import { QUERY_STALE_TIMES } from '@/hooks/queries/query-constants'
 import {
-  itemsKeys,
-  seriesKeys,
+  itemsQueryOptions,
+  seriesQueryOptions,
   useItem,
   useSeasons,
-} from '@/hooks/queries/use-items'
+} from '@/services/items/queries'
 import { Skeleton } from '@/components/ui/skeleton'
 import { LightRays } from '@/components/ui/light-rays'
 import { RouteErrorFallback } from '@/components/ui/route-error-fallback'
 import { FeatureErrorBoundary } from '@/components/ui/feature-error-boundary'
-import { getItemById, getSeasons } from '@/services/items/api'
 import { getBestImageUrl } from '@/services/video/api'
 import { useVibrantColor } from '@/hooks/use-vibrant-color'
 
@@ -89,16 +87,8 @@ export const Route = createFileRoute('/series/$itemId')({
 
     // Prefetch series and seasons data in parallel
     await Promise.all([
-      queryClient.ensureQueryData({
-        queryKey: itemsKeys.detail(itemId),
-        queryFn: () => getItemById(itemId),
-        staleTime: QUERY_STALE_TIMES.LONG,
-      }),
-      queryClient.ensureQueryData({
-        queryKey: seriesKeys.seasons(itemId),
-        queryFn: () => getSeasons(itemId),
-        staleTime: QUERY_STALE_TIMES.LONG,
-      }),
+      queryClient.ensureQueryData(itemsQueryOptions.detail(itemId)),
+      queryClient.ensureQueryData(seriesQueryOptions.seasons(itemId)),
     ])
   },
   onError: () => {
